face_emotion/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-dotenv
requests
